Switch to using an abstract base class for DB connection [ch1232]
